Remove compiler warning
git-svn-id: http://miranda.googlecode.com/svn/trunk@4394 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/srmm/statusicon.c
@@ -104,7 +104,7 @@ int ModifyStatusIcon(WPARAM wParam, LPARAM lParam) {
 				char buff[256];
 				HWND hwnd;
 				sprintf(buff, "SRMMStatusIconFlags%d", sid->dwId);
-				DBWriteContactSettingByte(hContact, sid->szModule, buff, sid->flags);
+				DBWriteContactSettingByte(hContact, sid->szModule, buff, (BYTE)sid->flags);
 				if (hwnd = WindowList_Find(g_dat->hMessageWindowList, hContact)) {
 					PostMessage(hwnd, DM_STATUSICONCHANGE, 0, 0);
 				}
